Clean some code
